gae_human/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests
google-cloud-logging
numpy==1.21.6
tensorflow==2.9.1